+++ src/ldc/supervised/classification/__init__.py
@@ -1,4 +1,5 @@
 from ._core import ClassificationData, ClassificationFilter, ClassificationReader, BatchClassificationWriter, StreamClassificationWriter
+from ._csv import CsvClassificationReader, CsvClassificationWriter, TsvClassificationReader, TsvClassificationWriter
 from ._jsonlines import JsonLinesClassificationReader, JsonLinesClassificationWriter
 from ._classification_label_map import ClassificationLabelMap
 from ._parquet import ParquetClassificationReader, ParquetClassificationWriter